Fixing warnings

+++ Rhombus/src/Rhombus/Core/Application.cpp
@@ -32,7 +32,7 @@ namespace rhombus {
 
 		m_Window = std::unique_ptr<Window>(Window::Create(WindowParams(m_Specification.name)));
 		m_Window->SetEventCallback(BIND_EVENT_FN(OnEvent));
-		SetViewport(0.0f, 0.0f, m_Window->GetWidth(), m_Window->GetHeight());
+		SetViewport(0.0f, 0.0f, (float)m_Window->GetWidth(), (float)m_Window->GetHeight());
 
 		Renderer::Init();
 		ScriptEngine::Init();
